+++ external_storage_openerp/attachment/attachment.py
@@ -37,5 +37,13 @@ class ir_attachment(osv.osv):
     _inherit = 'ir.attachment'
     _columns = {'db_datas': fields.binary('Database Data', store='s3'),
                 }
+
+    #Unlink Method to delete related records of attachements from
+    #AWS S3 Lookup table of openerp
+    def unlink(self, cr, uid, ids, context=None):
+        lookup_obj = self.pool.get('lookup')
+        lookup_ids = lookup_obj.search(cr, uid, [('res_id', 'in', ids)])
+        lookup_obj.unlink(cr, uid, lookup_ids, context=context)
+        return super(ir_attachment, self).unlink(cr, uid, ids, context=context)
 ir_attachment()
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4: