+++ doc/release_notes.rst
@@ -1,6 +1,17 @@
 Release Notes
 =============
 
+v0.3
+----
+:Date: 2013/09/13
+
+* Create logic for unique filename in 'lookup' object of openerp,
+  use same file name to store in AWS S3 bucket.
+* Added functionality for deleting files; if attached files are deleted
+  from openerp 'lookup' object then respective file with same name
+  will be deleted from AWS S3 bucket 
+
+
 v0.2
 ----
 :Date: 2013/08/21